feat: 기존 ready 상태 추상 클래스 분기 기능 작성
세부사항
- 기존 Ready 상태 클래스 추상화
- UserReady 상태 기능 구현
- DealerReady 상태 기능 구현
