+++ src/main/java/domain/state/Ready.java
@@ -3,24 +3,9 @@ package domain.state;
 import domain.card.Card;
 import domain.card.Hand;
 
-public final class Ready extends State {
-    public Ready() {
-        super(new Hand());
-    }
-
+public abstract class Ready extends State {
     public Ready(Hand hand) {
         super(hand);
-    }
-
-    @Override
-    public State draw(Card card) {
-        final Hand newHand = getHand().add(card);
-
-        if (getHand().isEmpty())
-            return new Ready(newHand);
-        if (newHand.isBlackjack())
-            return new Blackjack(newHand);
-        return new Hit(newHand);
     }
 
     @Override
